boto_elbv2 fix
