+++ salt/cloud/__init__.py
@@ -33,6 +33,7 @@ import salt.utils
 import salt.utils.args
 import salt.utils.cloud
 import salt.utils.context
+import salt.utils.crypt
 import salt.utils.dictupdate
 import salt.utils.files
 import salt.syspaths
@@ -2300,7 +2301,7 @@ def create_multiprocessing(parallel_data, queue=None):
     This function will be called from another process when running a map in
     parallel mode. The result from the create is always a json object.
     '''
-    salt.utils.reinit_crypto()
+    salt.utils.crypt.reinit_crypto()
 
     parallel_data['opts']['output'] = 'json'
     cloud = Cloud(parallel_data['opts'])
@@ -2332,7 +2333,7 @@ def destroy_multiprocessing(parallel_data, queue=None):
     This function will be called from another process when running a map in
     parallel mode. The result from the destroy is always a json object.
     '''
-    salt.utils.reinit_crypto()
+    salt.utils.crypt.reinit_crypto()
 
     parallel_data['opts']['output'] = 'json'
     clouds = salt.loader.clouds(parallel_data['opts'])
@@ -2368,7 +2369,7 @@ def run_parallel_map_providers_query(data, queue=None):
     This function will be called from another process when building the
     providers map.
     '''
-    salt.utils.reinit_crypto()
+    salt.utils.crypt.reinit_crypto()
 
     cloud = Cloud(data['opts'])
     try:
