+++ salt/config/__init__.py
@@ -24,7 +24,6 @@ from salt.ext.six.moves.urllib.parse import urlparse
 # pylint: enable=import-error,no-name-in-module
 
 # Import salt libs
-import salt.utils  # Can be removed once is_dictlist, ip_bracket are moved
 import salt.utils.data
 import salt.utils.dictupdate
 import salt.utils.files
@@ -36,6 +35,7 @@ import salt.utils.user
 import salt.utils.validate.path
 import salt.utils.xdg
 import salt.utils.yamlloader as yamlloader
+import salt.utils.zeromq
 import salt.syspaths
 import salt.exceptions
 from salt.utils.locales import sdecode
@@ -3915,7 +3915,7 @@ def client_config(path, env_var='SALT_CLIENT_CONFIG', defaults=None):
     # Make sure the master_uri is set
     if 'master_uri' not in opts:
         opts['master_uri'] = 'tcp://{ip}:{port}'.format(
-            ip=salt.utils.ip_bracket(opts['interface']),
+            ip=salt.utils.zeromq.ip_bracket(opts['interface']),
             port=opts['ret_port']
         )
