+++ salt/crypt.py
@@ -50,7 +50,8 @@ import salt.defaults.exitcodes
 import salt.payload
 import salt.transport.client
 import salt.transport.frame
-import salt.utils  # Can be removed when pem_finger, reinit_crypto are moved
+import salt.utils  # Can be removed when pem_finger is moved
+import salt.utils.crypt
 import salt.utils.decorators
 import salt.utils.event
 import salt.utils.files
@@ -113,7 +114,7 @@ def gen_keys(keydir, keyname, keysize, user=None, passphrase=None):
     priv = u'{0}.pem'.format(base)
     pub = u'{0}.pub'.format(base)
 
-    salt.utils.reinit_crypto()
+    salt.utils.crypt.reinit_crypto()
     gen = RSA.generate(bits=keysize, e=65537)
     if os.path.isfile(priv):
         # Between first checking and the generation another process has made
@@ -446,7 +447,7 @@ class AsyncAuth(object):
 
         self.io_loop = io_loop or tornado.ioloop.IOLoop.current()
 
-        salt.utils.reinit_crypto()
+        salt.utils.crypt.reinit_crypto()
         key = self.__key(self.opts)
         # TODO: if we already have creds for this key, lets just re-use
         if key in AsyncAuth.creds_map:
