+++ salt/minion.py
@@ -98,6 +98,7 @@ import salt.utils.minion
 import salt.utils.minions
 import salt.utils.network
 import salt.utils.platform
+import salt.utils.process
 import salt.utils.schedule
 import salt.utils.user
 import salt.utils.zeromq
@@ -156,8 +157,11 @@ def resolve_dns(opts, fallback=True):
         try:
             if opts[u'master'] == u'':
                 raise SaltSystemExit
-            ret[u'master_ip'] = \
-                    salt.utils.dns_check(opts[u'master'], int(opts[u'master_port']), True, opts[u'ipv6'])
+            ret[u'master_ip'] = salt.utils.network.dns_check(
+                opts[u'master'],
+                int(opts[u'master_port']),
+                True,
+                opts[u'ipv6'])
         except SaltClientError:
             if opts[u'retry_dns']:
                 while True:
@@ -170,9 +174,11 @@ def resolve_dns(opts, fallback=True):
                         print(u'WARNING: {0}'.format(msg))
                     time.sleep(opts[u'retry_dns'])
                     try:
-                        ret[u'master_ip'] = salt.utils.dns_check(
-                            opts[u'master'], int(opts[u'master_port']), True, opts[u'ipv6']
-                        )
+                        ret[u'master_ip'] = salt.utils.network.dns_check(
+                            opts[u'master'],
+                            int(opts[u'master_port']),
+                            True,
+                            opts[u'ipv6'])
                         break
                     except SaltClientError:
                         pass
@@ -1470,12 +1476,12 @@ class Minion(MinionBase):
             # Shutdown the multiprocessing before daemonizing
             salt.log.setup.shutdown_multiprocessing_logging()
 
-            salt.utils.daemonize_if(opts)
+            salt.utils.process.daemonize_if(opts)
 
             # Reconfigure multiprocessing logging after daemonizing
             salt.log.setup.setup_multiprocessing_logging()
 
-        salt.utils.appendproctitle(u'{0}._thread_return {1}'.format(cls.__name__, data[u'jid']))
+        salt.utils.process.appendproctitle(u'{0}._thread_return {1}'.format(cls.__name__, data[u'jid']))
 
         sdata = {u'pid': os.getpid()}
         sdata.update(data)
@@ -1654,7 +1660,7 @@ class Minion(MinionBase):
         This method should be used as a threading target, start the actual
         minion side execution.
         '''
-        salt.utils.appendproctitle(u'{0}._thread_multi_return {1}'.format(cls.__name__, data[u'jid']))
+        salt.utils.process.appendproctitle(u'{0}._thread_multi_return {1}'.format(cls.__name__, data[u'jid']))
         multifunc_ordered = opts.get(u'multifunc_ordered', False)
         num_funcs = len(data[u'fun'])
         if multifunc_ordered:
