+++ salt/modules/inspectlib/collector.py
@@ -28,7 +28,7 @@ from salt.modules.inspectlib import kiwiproc
 from salt.modules.inspectlib.entities import (AllowedDir, IgnoredDir, Package,
                                               PayloadFile, PackageCfgFile)
 
-import salt.utils  # Can be removed when reinit_crypto is moved
+import salt.utils.crypt
 import salt.utils.files
 import salt.utils.fsutils
 import salt.utils.path
@@ -505,10 +505,10 @@ if __name__ == '__main__':
     # Double-fork stuff
     try:
         if os.fork() > 0:
-            salt.utils.reinit_crypto()
+            salt.utils.crypt.reinit_crypto()
             sys.exit(0)
         else:
-            salt.utils.reinit_crypto()
+            salt.utils.crypt.reinit_crypto()
     except OSError as ex:
         sys.exit(1)
 
@@ -518,12 +518,12 @@ if __name__ == '__main__':
     try:
         pid = os.fork()
         if pid > 0:
-            salt.utils.reinit_crypto()
+            salt.utils.crypt.reinit_crypto()
             with salt.utils.files.fopen(os.path.join(pidfile, EnvLoader.PID_FILE), 'w') as fp_:
                 fp_.write('{0}\n'.format(pid))
             sys.exit(0)
     except OSError as ex:
         sys.exit(1)
 
-    salt.utils.reinit_crypto()
+    salt.utils.crypt.reinit_crypto()
     main(dbfile, pidfile, mode)
