+++ salt/modules/network.py
@@ -57,7 +57,7 @@ def wol(mac, bcast='255.255.255.255', destport=9):
         salt '*' network.wol 080027136977 255.255.255.255 7
         salt '*' network.wol 08:00:27:13:69:77 255.255.255.255 7
     '''
-    dest = salt.utils.mac_str_to_bytes(mac)
+    dest = salt.utils.network.mac_str_to_bytes(mac)
     sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
     sock.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
     sock.sendto(b'\xff' * 6 + dest * 16, (bcast, int(destport)))
