+++ salt/runners/network.py
@@ -10,8 +10,8 @@ import logging
 import socket
 
 # Import salt libs
-import salt.utils
 import salt.utils.files
+import salt.utils.network
 
 log = logging.getLogger(__name__)
 
@@ -54,7 +54,7 @@ def wol(mac, bcast='255.255.255.255', destport=9):
         salt-run network.wol 080027136977 255.255.255.255 7
         salt-run network.wol 08:00:27:13:69:77 255.255.255.255 7
     '''
-    dest = salt.utils.mac_str_to_bytes(mac)
+    dest = salt.utils.network.mac_str_to_bytes(mac)
     sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
     sock.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
     sock.sendto(b'\xff' * 6 + dest * 16, (bcast, int(destport)))
