+++ salt/utils/vt.py
@@ -52,7 +52,7 @@ except ImportError:
     import resource
 
 # Import salt libs
-import salt.utils
+import salt.utils.crypt
 import salt.utils.stringutils
 from salt.ext.six import string_types
 from salt.log.setup import LOG_LEVELS
@@ -493,7 +493,7 @@ class Terminal(object):
                 # Close parent FDs
                 os.close(stdout_parent_fd)
                 os.close(stderr_parent_fd)
-                salt.utils.reinit_crypto()
+                salt.utils.crypt.reinit_crypto()
 
                 # ----- Make STDOUT the controlling PTY --------------------->
                 child_name = os.ttyname(stdout_child_fd)
@@ -554,7 +554,7 @@ class Terminal(object):
                 # <---- Duplicate Descriptors --------------------------------
             else:
                 # Parent. Close Child PTY's
-                salt.utils.reinit_crypto()
+                salt.utils.crypt.reinit_crypto()
                 os.close(stdout_child_fd)
                 os.close(stderr_child_fd)
