+++ salt/utils/zeromq.py
@@ -30,3 +30,13 @@ def check_ipc_path_max_len(uri):
                 uri, ipc_path_max_len
             )
         )
+
+
+def ip_bracket(addr):
+    '''
+    Convert IP address representation to ZMQ (URL) format. ZMQ expects
+    brackets around IPv6 literals, since they are used in URLs.
+    '''
+    if addr and ':' in addr and not addr.startswith('['):
+        return '[{0}]'.format(addr)
+    return addr
